fix build error

+++ pkg/plugins/dynamic/plugins.go
@@ -90,11 +90,11 @@ func (ds *DynamicScheduler) Score(ctx context.Context, state *framework.CycleSta
 
 	finalScore := score - int(hotValue*10)
 
-	finalScore = utils.NormalizeScore(finalScore,framework.MaxNodeScore,framework.MinNodeScore)
+	finalScore = utils.NormalizeScore(int64(finalScore),framework.MaxNodeScore,framework.MinNodeScore)
 
 	klog.V(4).Infof("[crane] Node[%s]'s final score is %d, while score is %d and hot value is %f", node.Name, finalScore, score, hotValue)
 
-	return int64(finalScore), nil
+	return finalScore, nil
 }
 
 func (ds *DynamicScheduler) ScoreExtensions() framework.ScoreExtensions {
